Refactor composables for testing home, movie details, search
